+++ vortex-array/src/stream/ext.rs
@@ -12,7 +12,7 @@ pub trait ArrayStreamExt: ArrayStream {
     /// Collect the stream into a single `Array`.
     ///
     /// If the stream yields multiple chunks, they will be returned as a [`ChunkedArray`].
-    fn into_array_data(self) -> impl Future<Output = VortexResult<Array>>
+    fn into_array(self) -> impl Future<Output = VortexResult<Array>>
     where
         Self: Sized,
     {
