+++ vortex-file/src/lib.rs
@@ -22,9 +22,9 @@
 //!
 //! # Reading
 //!
-//! Reading is implemented by [`VortexFile`]. It's "opened" by [`VortexOpenOptions`], which can be provided with information about's the file's
+//! Reading is implemented by [`FileVortexFile`]. It's "opened" by [`VortexOpenOptions`], which can be provided with information about's the file's
 //! structure to save on IO before the actual data read. Once the file is open and has done the initial IO work to understand its own structure,
-//! it can be turned into a stream by calling [`VortexFile::scan`] with a [`Scan`], which defines filtering and projection on the file.
+//! it can be turned into a stream by calling [`FileVortexFile::scan`] with a [`Scan`], which defines filtering and projection on the file.
 //!
 //! The file manages IO-oriented work and CPU-oriented work on two different underlying runtimes, which are configurable and pluggable with multiple provided implementations (Tokio, Rayon etc.).
 //! It also caches buffers between stages of the scan, saving on duplicate IO. The cache can also be reused between scans of the same file (See [`SegmentCache`](`crate::segments::SegmentCache`)).
@@ -91,20 +91,23 @@
 //! If you ultimately seek Arrow arrays, [`VortexRecordBatchReader`][`crate::read::VortexRecordBatchReader`] converts an open
 //! Vortex file into a [`RecordBatchReader`](arrow_array::RecordBatchReader).
 
-mod exec;
+pub mod exec;
 mod file;
 mod footer;
-pub mod io;
+mod memory;
 mod open;
 pub mod read;
 pub mod segments;
+mod split_by;
 #[cfg(test)]
 mod tests;
+mod unified;
 mod writer;
 
 pub use file::*;
 pub use footer::{FileLayout, Segment};
 pub use forever_constant::*;
+pub use memory::*;
 pub use open::*;
 pub use writer::*;
