+++ vortex-file/src/open.rs
@@ -1,68 +1,78 @@
-mod exec;
-mod split_by;
-
 use std::sync::Arc;
 
-pub use exec::*;
 use flatbuffers::root;
 use futures_util::stream::FuturesUnordered;
 use futures_util::{stream, StreamExt, TryStreamExt};
 use itertools::Itertools;
 use moka::future::CacheBuilder;
-pub use split_by::*;
 use vortex_array::ContextRef;
 use vortex_buffer::{ByteBuffer, ByteBufferMut};
 use vortex_dtype::DType;
 use vortex_error::{vortex_bail, vortex_err, VortexExpect, VortexResult};
 use vortex_flatbuffers::{dtype as fbd, footer as fb, FlatBuffer, ReadFlatBuffer};
 use vortex_io::VortexReadAt;
+use vortex_layout::scan::{Scan, ScanDriver};
 use vortex_layout::segments::SegmentId;
 use vortex_layout::{Layout, LayoutContextRef, LayoutId};
+use vortex_sampling_compressor::ALL_ENCODINGS_CONTEXT;
 
 use crate::footer::{FileLayout, Postscript, Segment};
-use crate::io::file::FileIoDriver;
 use crate::segments::{InMemorySegmentCache, NoOpSegmentCache, SegmentCache};
-use crate::{VortexFile, EOF_SIZE, MAGIC_BYTES, VERSION};
+use crate::{FileVortexFile, InMemoryVortexFile, EOF_SIZE, MAGIC_BYTES, VERSION};
 
-const INITIAL_READ_SIZE: u64 = 1 << 20; // 1 MB
+/// Trait for a Vortex file.
+///
+/// This exists so different Vortex file types can configure their scan operations differently.
+pub struct VortexFile<F: VortexFileOpener> {
+    ctx: ContextRef,
+    file_layout: FileLayout,
+    file: F,
+}
+
+impl<F: VortexFileOpener> VortexFile<F> {
+    ///
+    pub fn scan(self) -> Scan<F::ScanDriver> {
+        let layout = self.file_layout.root_layout().clone();
+        let ctx = self.ctx.clone();
+        Scan::new(self.file, layout, ctx)
+    }
+}
+
+pub trait VortexFileOpener: Sized {
+    type Options;
+    type Read: VortexReadAt;
+    type ScanDriver: ScanDriver;
+
+    fn open(
+        ctx: ContextRef,
+        file_layout: FileLayout,
+        segment_cache: Arc<dyn SegmentCache>,
+        options: Self::Options,
+        read: Self::Read,
+    ) -> VortexResult<Self>;
+}
 
 /// Open options for a Vortex file reader.
-pub struct VortexOpenOptions {
+pub struct VortexOpenOptions<F: VortexFileOpener> {
     /// The Vortex Array encoding context.
-    pub(crate) ctx: ContextRef,
+    ctx: ContextRef,
     /// The Vortex Layout encoding context.
     layout_ctx: LayoutContextRef,
     /// An optional, externally provided, file size.
     file_size: Option<u64>,
     /// An optional, externally provided, file layout.
+    // TODO(ngates): add an optional DType so we only read the layout segment.
     file_layout: Option<FileLayout>,
-    // TODO(ngates): also support a messages_middleware that can wrap a message cache to provide
-    //  additional caching, metrics, or other intercepts, etc. It should support synchronous
-    //  read + write of Map<MessageId, ByteBuffer> or similar.
+    segment_cache: Arc<dyn SegmentCache>,
     initial_read_size: u64,
-    split_by: SplitBy,
-    segment_cache: Option<Arc<dyn SegmentCache>>,
-    execution_mode: Option<ExecutionMode>,
-    // TODO(ngates): allow fully configurable I/O driver.
-    io_concurrency: usize,
-    exec_concurrency: Option<usize>,
+    options: F::Options,
 }
 
-impl VortexOpenOptions {
-    pub fn new(ctx: ContextRef) -> Self {
-        Self {
-            ctx,
-            layout_ctx: LayoutContextRef::default(),
-            file_size: None,
-            file_layout: None,
-            initial_read_size: INITIAL_READ_SIZE,
-            split_by: SplitBy::Layout,
-            segment_cache: None,
-            execution_mode: None,
-            // TODO(ngates): pick some numbers...
-            io_concurrency: 10,
-            exec_concurrency: None,
-        }
+impl<F: VortexFileOpener> VortexOpenOptions<F> {
+    /// Configure a Vortex Array context.
+    pub fn with_ctx(mut self, ctx: ContextRef) -> Self {
+        self.ctx = ctx;
+        self
     }
 
     /// Configure a layout context.
@@ -74,7 +84,7 @@ impl VortexOpenOptions {
     /// Configure a known file layout.
     ///
     /// If this is provided, then the Vortex file can be opened without performing any I/O.
-    /// Once open, the [`FileLayout`] can be accessed via [`VortexFile::file_layout`].
+    /// Once open, the [`FileLayout`] can be accessed via [`FileVortexFile::file_layout`].
     pub fn with_file_layout(mut self, file_layout: FileLayout) -> Self {
         self.file_layout = Some(file_layout);
         self
@@ -98,17 +108,9 @@ impl VortexOpenOptions {
         Ok(self)
     }
 
-    /// Configure how to split the file into batches for reading.
-    ///
-    /// Defaults to [`SplitBy::Layout`].
-    pub fn with_split_by(mut self, split_by: SplitBy) -> Self {
-        self.split_by = split_by;
-        self
-    }
-
     /// Configure a custom [`SegmentCache`].
     pub fn with_segment_cache(mut self, segment_cache: Arc<dyn SegmentCache>) -> Self {
-        self.segment_cache = Some(segment_cache);
+        self.segment_cache = segment_cache;
         self
     }
 
@@ -116,79 +118,67 @@ impl VortexOpenOptions {
     pub fn without_segment_cache(self) -> Self {
         self.with_segment_cache(Arc::new(NoOpSegmentCache))
     }
+}
 
-    /// Configure the execution mode
-    pub fn with_execution_mode(mut self, execution_mode: ExecutionMode) -> Self {
-        self.execution_mode = Some(execution_mode);
-        self
-    }
-
-    /// Configure the number of concurrent I/O requests.
-    pub fn with_io_concurrency(mut self, io_concurrency: usize) -> Self {
-        self.io_concurrency = io_concurrency;
-        self
-    }
-
-    /// Override the default split-by concurrency.
-    ///
-    /// It is recommended to use more split-by concurrency than I/O concurrency to ensure there
-    /// are always I/O operations enqueued.
-    pub fn with_exec_concurrency(mut self, exec_concurrency: usize) -> Self {
-        self.exec_concurrency = Some(exec_concurrency);
-        self
+impl VortexOpenOptions<InMemoryVortexFile> {
+    /// Open an in-memory file contained in the provided buffer.
+    pub fn in_memory() -> Self {
+        Self {
+            ctx: ALL_ENCODINGS_CONTEXT.clone(),
+            layout_ctx: Arc::new(Default::default()),
+            file_size: None,
+            file_layout: None,
+            segment_cache: Arc::new(NoOpSegmentCache),
+            initial_read_size: 0,
+            options: (),
+        }
     }
 }
 
-impl VortexOpenOptions {
-    /// Open the Vortex file using asynchronous IO.
-    pub async fn open<R: VortexReadAt>(
-        mut self,
-        read: R,
-    ) -> VortexResult<VortexFile<FileIoDriver<R>>> {
-        // Set up our segment cache.
-        let segment_cache = self.segment_cache.as_ref().cloned().unwrap_or_else(|| {
-            Arc::new(InMemorySegmentCache::new(
+pub struct FileOpenOptions;
+
+impl<R: VortexReadAt> VortexOpenOptions<FileVortexFile<R>> {
+    const INITIAL_READ_SIZE: u64 = 1 << 20; // 1 MB
+
+    pub fn file() -> Self {
+        Self {
+            // TODO(ngates): move this context into the vortex-file crate
+            ctx: ALL_ENCODINGS_CONTEXT.clone(),
+            layout_ctx: LayoutContextRef::default(),
+            file_size: None,
+            file_layout: None,
+            segment_cache: Arc::new(InMemorySegmentCache::new(
                 // For now, use a fixed 1GB overhead.
                 CacheBuilder::new(1 << 30),
-            ))
-        });
+            )),
+            initial_read_size: Self::INITIAL_READ_SIZE,
+            options: FileOpenOptions,
+        }
+    }
+}
+
+impl<F: VortexFileOpener> VortexOpenOptions<F> {
+    /// Open the Vortex file using asynchronous IO.
+    pub async fn open<R: Into<F::Read>>(mut self, read: R) -> VortexResult<VortexFile<F>> {
+        let read = read.into();
 
         // If we need to read the file layout, then do so.
         let file_layout = match self.file_layout.take() {
-            None => self.read_file_layout(&read, segment_cache.as_ref()).await?,
+            None => self.read_file_layout(&read).await?,
             Some(file_layout) => file_layout,
         };
 
-        // Set up the I/O driver.
-        let io_driver = FileIoDriver {
-            read,
-            file_layout: file_layout.clone(),
-            concurrency: self.io_concurrency,
-            segment_cache,
-        };
-
-        // Set up the execution driver.
-        let exec_driver = self
-            .execution_mode
-            .unwrap_or_default()
-            .into_driver(self.exec_concurrency.unwrap_or(self.io_concurrency * 2));
-
-        // Finally, create the VortexFile.
-        Ok(VortexFile {
-            ctx: self.ctx.clone(),
+        F::open(
+            self.ctx,
             file_layout,
-            io_driver,
-            exec_driver,
-            split_by: self.split_by,
-        })
+            self.segment_cache,
+            self.options,
+            read,
+        )
     }
 
     /// Read the [`FileLayout`] from the file.
-    pub(crate) async fn read_file_layout<R: VortexReadAt>(
-        &self,
-        read: &R,
-        segment_cache: &dyn SegmentCache,
-    ) -> VortexResult<FileLayout> {
+    async fn read_file_layout<R: VortexReadAt>(&self, read: &R) -> VortexResult<FileLayout> {
         // Fetch the file size and perform the initial read.
         let file_size = match self.file_size {
             None => read.size().await?,
@@ -239,7 +229,7 @@ impl VortexOpenOptions {
 
         // If the initial read happened to cover any segments, then we can populate the
         // segment cache
-        self.populate_segments(initial_offset, &initial_read, &file_layout, segment_cache)
+        self.populate_segments(initial_offset, &initial_read, &file_layout)
             .await?;
 
         Ok(file_layout)
@@ -339,7 +329,6 @@ impl VortexOpenOptions {
         initial_offset: u64,
         initial_read: &ByteBuffer,
         file_layout: &FileLayout,
-        segments: &dyn SegmentCache,
     ) -> VortexResult<()> {
         stream::iter(
             file_layout
@@ -354,7 +343,7 @@ impl VortexOpenOptions {
                         .slice(offset..offset + (segment.length as usize))
                         .aligned(segment.alignment);
 
-                    segments.put(segment_id, buffer).await
+                    self.segment_cache.put(segment_id, buffer).await
                 }),
         )
         .collect::<FuturesUnordered<_>>()
